+++ Sources/Views/ALKChatCell.swift
@@ -298,7 +298,7 @@ public final class ALKChatCell: MGSwipeTableCell, Localizable {
         deleteButton.tintColor = .white
         deleteButton.accessibilityIdentifier = "SwippableDeleteIcon"
         deleteButton.frame = CGRect.init(x: 0, y: 0, width: 69, height: 69)
-        if !viewModel.isGroupChat {
+        if (!viewModel.isGroupChat || (viewModel.channelKey != nil && ALChannelService().isChannelLeft(viewModel.channelKey))) {
             let leaveTitle = localizedString(forKey: "DeleteButtonName", withDefaultValue: SystemMessage.ButtonName.Delete, fileName: localizationFileName)
             deleteButton.setTitle(leaveTitle, for: .normal)
         } else {