+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 2.5.1
 
 ### Enhancments
-- [AL-3490]Added support for Back button action in Conversation VC if the VC is presented not pushed.
+- [AL-3490] Added support for Back button action in Conversation VC if the VC is presented not pushed.
+- [AL-3397] Added support for sharing contacts
 
 ### Fixes
 - [AL-3486]Fixed an issue where in some cases view was in an incorrect state if the keyboard is visible.